refactor: comprehensive code review and refactoring
- Sync shared constants with all 21 Prisma enums (PlanType, ProductStatus, etc.)
- Add 14 missing type exports to shared package
- Add cascade delete rules to 11 parent-child relationships
- Add compound indexes for performance (Product, Contract, NotificationLog)
- Extract duplicated number generation to shared sequence.util.ts
- Replace magic numbers with named BUSINESS_RULES constants
- Add input validation to installment calculation utility
- Fix ProductStatus DTO to include all 14 statuses
- Split ProductsService (1,218 lines) into 3 focused services
- Standardize error handling: replace err.response?.data?.message with getErrorMessage()
- Add error UI states to all dashboard queries with retry buttons
- Add Thai labels for all enum types in shared STATUS_LABELS
