+++ apps/api/src/modules/products/dto/create-product.dto.ts
@@ -56,7 +56,7 @@ export class CreateProductDto {
   @IsString()
   branchId: string;
 
-  @IsIn(['PO_RECEIVED', 'INSPECTION', 'IN_STOCK', 'RESERVED', 'SOLD_INSTALLMENT', 'SOLD_CASH', 'REPOSSESSED', 'REFURBISHED', 'SOLD_RESELL'])
+  @IsIn(['PO_RECEIVED', 'QC_PENDING', 'PHOTO_PENDING', 'INSPECTION', 'IN_STOCK', 'RESERVED', 'SOLD_INSTALLMENT', 'SOLD_CASH', 'REPOSSESSED', 'REFURBISHED', 'SOLD_RESELL', 'DAMAGED', 'LOST', 'WRITTEN_OFF'])
   @IsOptional()
   status?: string;
