+++ apps/api/src/modules/products/products.controller.ts
@@ -1,5 +1,7 @@
 import { Controller, Get, Post, Patch, Delete, Param, Body, Query, UseGuards } from '@nestjs/common';
 import { ProductsService } from './products.service';
+import { ProductsPricingService } from './products-pricing.service';
+import { ProductsStockService } from './products-stock.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { CreateProductPriceDto, UpdateProductPriceDto } from './dto/product-price.dto';
@@ -12,7 +14,11 @@ import { CurrentUser } from '../auth/decorators/current-user.decorator';
 @Controller('products')
 @UseGuards(JwtAuthGuard, RolesGuard)
 export class ProductsController {
-  constructor(private productsService: ProductsService) {}
+  constructor(
+    private productsService: ProductsService,
+    private productsPricingService: ProductsPricingService,
+    private productsStockService: ProductsStockService,
+  ) {}
 
   @Get()
   findAll(
@@ -42,7 +48,7 @@ export class ProductsController {
     @Query('page') page?: string,
     @Query('limit') limit?: string,
   ) {
-    return this.productsService.getStock({
+    return this.productsStockService.getStock({
       search, branchId, status, category, brand,
       page: page ? parseInt(page) : undefined,
       limit: limit ? parseInt(limit) : undefined,
@@ -51,7 +57,7 @@ export class ProductsController {
 
   @Get('stock/dashboard')
   getStockDashboard(@Query('branchId') branchId?: string) {
-    return this.productsService.getStockDashboard(branchId);
+    return this.productsStockService.getStockDashboard(branchId);
   }
 
   @Get('brands')
@@ -64,7 +70,7 @@ export class ProductsController {
     @Query('days') days?: string,
     @Query('branchId') branchId?: string,
   ) {
-    return this.productsService.getWarrantyExpiring(
+    return this.productsStockService.getWarrantyExpiring(
       days ? parseInt(days) : 30,
       branchId,
     );
@@ -73,7 +79,7 @@ export class ProductsController {
   @Get('supplier-performance')
   @Roles('OWNER', 'BRANCH_MANAGER')
   getSupplierPerformance() {
-    return this.productsService.getSupplierPerformance();
+    return this.productsStockService.getSupplierPerformance();
   }
 
   // Static transfer routes MUST be before transfers/:transferId
@@ -81,13 +87,13 @@ export class ProductsController {
   @Get('transfers/pending')
   @Roles('OWNER', 'BRANCH_MANAGER')
   getPendingTransfers(@Query('branchId') branchId?: string) {
-    return this.productsService.getPendingTransfers(branchId);
+    return this.productsStockService.getPendingTransfers(branchId);
   }
 
   @Get('transfers/in-transit')
   @Roles('OWNER', 'BRANCH_MANAGER')
   getInTransitTransfers(@Query('branchId') branchId?: string) {
-    return this.productsService.getInTransitTransfers(branchId);
+    return this.productsStockService.getInTransitTransfers(branchId);
   }
 
   @Get('transfers/history')
@@ -100,7 +106,7 @@ export class ProductsController {
     @Query('page') page?: string,
     @Query('limit') limit?: string,
   ) {
-    return this.productsService.getTransferHistory({
+    return this.productsStockService.getTransferHistory({
       branchId, status, startDate, endDate,
       page: page ? parseInt(page) : undefined,
       limit: limit ? parseInt(limit) : undefined,
@@ -109,7 +115,7 @@ export class ProductsController {
 
   @Get('transfers/:transferId')
   getTransferById(@Param('transferId') transferId: string) {
-    return this.productsService.getTransferById(transferId);
+    return this.productsStockService.getTransferById(transferId);
   }
 
   @Get(':id/workflow')
@@ -145,7 +151,7 @@ export class ProductsController {
   @Post(':id/prices')
   @Roles('OWNER', 'BRANCH_MANAGER')
   addPrice(@Param('id') id: string, @Body() dto: CreateProductPriceDto) {
-    return this.productsService.addPrice(id, dto);
+    return this.productsPricingService.addPrice(id, dto);
   }
 
   @Patch(':id/prices/:priceId')
@@ -155,13 +161,13 @@ export class ProductsController {
     @Param('priceId') priceId: string,
     @Body() dto: UpdateProductPriceDto,
   ) {
-    return this.productsService.updatePrice(id, priceId, dto);
+    return this.productsPricingService.updatePrice(id, priceId, dto);
   }
 
   @Delete(':id/prices/:priceId')
   @Roles('OWNER', 'BRANCH_MANAGER')
   removePrice(@Param('id') id: string, @Param('priceId') priceId: string) {
-    return this.productsService.removePrice(id, priceId);
+    return this.productsPricingService.removePrice(id, priceId);
   }
 
   // === Reservation Endpoints ===
@@ -169,13 +175,13 @@ export class ProductsController {
   @Post(':id/reserve')
   @Roles('OWNER', 'BRANCH_MANAGER', 'SALES')
   reserve(@Param('id') id: string, @Body('reason') reason?: string) {
-    return this.productsService.reserve(id, reason);
+    return this.productsStockService.reserve(id, reason);
   }
 
   @Post(':id/unreserve')
   @Roles('OWNER', 'BRANCH_MANAGER', 'SALES')
   unreserve(@Param('id') id: string) {
-    return this.productsService.unreserve(id);
+    return this.productsStockService.unreserve(id);
   }
 
   // === Transfer Endpoints ===
@@ -186,7 +192,7 @@ export class ProductsController {
     @Body() dto: BulkTransferDto,
     @CurrentUser() user: { id: string },
   ) {
-    return this.productsService.bulkTransfer(dto, user.id);
+    return this.productsStockService.bulkTransfer(dto, user.id);
   }
 
   @Post(':id/transfer')
@@ -196,7 +202,7 @@ export class ProductsController {
     @Body() dto: TransferProductDto,
     @CurrentUser() user: { id: string },
   ) {
-    return this.productsService.transfer(id, dto, user.id);
+    return this.productsStockService.transfer(id, dto, user.id);
   }
 
   @Post('transfers/:transferId/dispatch')
@@ -206,7 +212,7 @@ export class ProductsController {
     @Body() dto: DispatchTransferDto,
     @CurrentUser() user: { id: string },
   ) {
-    return this.productsService.dispatchTransfer(transferId, user.id, dto.trackingNote);
+    return this.productsStockService.dispatchTransfer(transferId, user.id, dto.trackingNote);
   }
 
   @Post('transfers/:transferId/confirm')
@@ -215,7 +221,7 @@ export class ProductsController {
     @Param('transferId') transferId: string,
     @CurrentUser() user: { id: string },
   ) {
-    return this.productsService.confirmTransfer(transferId, user.id);
+    return this.productsStockService.confirmTransfer(transferId, user.id);
   }
 
   @Post('transfers/:transferId/reject')
@@ -225,7 +231,7 @@ export class ProductsController {
     @CurrentUser() user: { id: string },
     @Body('reason') reason?: string,
   ) {
-    return this.productsService.rejectTransfer(transferId, user.id, reason);
+    return this.productsStockService.rejectTransfer(transferId, user.id, reason);
   }
 
 }
