+++ apps/api/src/modules/products/products.module.ts
@@ -1,10 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ProductsController } from './products.controller';
 import { ProductsService } from './products.service';
+import { ProductsPricingService } from './products-pricing.service';
+import { ProductsStockService } from './products-stock.service';
 
 @Module({
   controllers: [ProductsController],
-  providers: [ProductsService],
-  exports: [ProductsService],
+  providers: [ProductsService, ProductsPricingService, ProductsStockService],
+  exports: [ProductsService, ProductsPricingService, ProductsStockService],
 })
 export class ProductsModule {}
