+++ apps/web/src/pages/UsersPage.tsx
@@ -59,8 +59,7 @@ export default function UsersPage() {
       toast.success(editingUser ? 'แก้ไขผู้ใช้สำเร็จ' : 'เพิ่มผู้ใช้สำเร็จ');
       closeModal();
     },
-    onError: (err: { response?: { data?: { message?: string } } }) =>
-      toast.error(err.response?.data?.message || 'เกิดข้อผิดพลาด'),
+    onError: (err: unknown) => toast.error(getErrorMessage(err)),
   });
 
   const toggleActiveMutation = useMutation({
